requirements.txt
python-dotenv
requests
streamlit
huggingface-hub
sentence-transformers
chromadb
numpy

